feito teste de criacao de usuario, automaticamente e criado um perfil de jogador associado ao usuario
feito teste de rota protegida com o token em questao

+++ src/main/resources/db/migration/V6__create_table_digimon.sql
@@ -1,6 +1,6 @@
 CREATE TABLE digimon (
-    id SERIAL PRIMARY KEY,
-    jogador_id INT NOT NULL REFERENCES jogador(id) ON DELETE CASCADE,
+    id BIGSERIAL PRIMARY KEY,
+    jogador_id BIGINT NOT NULL REFERENCES jogador(id) ON DELETE CASCADE,
     nome VARCHAR(60) NOT NULL,
     especie VARCHAR(60) NOT NULL,
     nivel INT DEFAULT 1,
